+++ packages/ui-transfer/src/components/quantityInput/index.tsx
@@ -11,6 +11,7 @@ import {
   Flex,
   inputFormatter,
   Spinner,
+  InputFormatter,
 } from "@orderly.network/ui";
 import { InputStatus } from "../../types";
 import { TokenOption } from "./tokenOption";
@@ -26,6 +27,7 @@ export type QuantityInputProps = {
   fetchBalance?: (token: string, decimals: number) => Promise<any>;
   loading?: boolean;
   testId?: string;
+  formatters?: InputFormatter[];
 } & Omit<InputProps, "onClear" | "suffix" | "onValueChange">;
 
 export const QuantityInput = forwardRef<HTMLInputElement, QuantityInputProps>(
@@ -43,6 +45,7 @@ export const QuantityInput = forwardRef<HTMLInputElement, QuantityInputProps>(
       onTokenChange,
       loading,
       placeholder,
+      formatters,
       ...rest
     } = props;
 
@@ -172,11 +175,13 @@ export const QuantityInput = forwardRef<HTMLInputElement, QuantityInputProps>(
           onValueChange={(value) => {
             onValueChange?.(value);
           }}
-          formatters={[
-            inputFormatter.numberFormatter,
-            inputFormatter.dpFormatter(token?.precision ?? 2),
-            inputFormatter.currencyFormatter,
-          ]}
+          formatters={
+            formatters || [
+              inputFormatter.numberFormatter,
+              inputFormatter.dpFormatter(token?.precision ?? 2),
+              inputFormatter.currencyFormatter,
+            ]
+          }
           {...rest}
           classNames={{
             ...classNames,